Changing MOVs with immediate values to MOVW
Encodings "T2" and "T3" of MOV allow 12-bit and 16-bit immediate values respectively. The reference manual states that T2 is preferred whenever available, which causes problems (say for the inverses of 3 and 5 mod 2^16). Some newer compilers are able to check the size of the immediate value and use the appropriate coding at compile time, but to force T3 (if the compiler cannot distinguish) one must use MOVW.
MOVs copying between registers are unaffected.